Refactored ReplaceNodes into utils
